+++ decor/eta.go
@@ -98,7 +98,7 @@ func (d *movingAverageETA) Decor(st *Statistics) string {
 	return d.FormatMsg(str)
 }
 
-func (d *movingAverageETA) NextAmount(n int, wdd ...time.Duration) {
+func (d *movingAverageETA) NextAmount(n int64, wdd ...time.Duration) {
 	var workDuration time.Duration
 	for _, wd := range wdd {
 		workDuration = wd
