+++ decor/speed.go
@@ -192,7 +192,7 @@ func (d *movingAverageSpeed) Decor(st *Statistics) string {
 	return d.FormatMsg(d.msg)
 }
 
-func (d *movingAverageSpeed) NextAmount(n int, wdd ...time.Duration) {
+func (d *movingAverageSpeed) NextAmount(n int64, wdd ...time.Duration) {
 	var workDuration time.Duration
 	for _, wd := range wdd {
 		workDuration = wd
